refactor(agentic): move corrector agent into its own directory
